// src/canisters/ledgers/ledger_ogy/ogy.did.d.ts
export interface BinaryAccountBalanceArgs {
    account: Array<number>;
}
// ...
export type Result_1 = { Ok: bigint } | { Err: TransferError_1 };
// export interface SendArgs {
//     to: string;
//     fee: Tokens;
//     memo: bigint;
//     from_subaccount: [] | [Array<number>];
//     created_at_time: [] | [TimeStamp];
//     amount: Tokens;
// }
// export interface StandardRecord {
//     url: string;
//     name: string;
// }
// export interface Symbol {
//     symbol: string;
// }
export interface TimeStamp {
    timestamp_nanos: bigint;
}
export interface Tokens {
    e8s: bigint;
}
// export interface TransferArg {
//     to: Account;
//     fee: [] | [bigint];
//     memo: [] | [Array<number>];
//     from_subaccount: [] | [Array<number>];
//     created_at_time: [] | [bigint];
//     amount: bigint;
// }
export interface TransferArgs {
    to: Array<number>;
    fee: Tokens;
    memo: bigint;
    from_subaccount: [] | [Array<number>];
    created_at_time: [] | [TimeStamp];
    amount: Tokens;
}
// ...
export type TransferError_1 =
    | {
          TxTooOld: { allowed_window_nanos: bigint };
      }
    | { BadFee: { expected_fee: Tokens } }
    | { TxDuplicate: { duplicate_of: bigint } }
    | { TxCreatedInFuture: null }
    | { InsufficientFunds: { balance: Tokens } };
// ...
export default interface _SERVICE {
    account_balance: (arg_0: BinaryAccountBalanceArgs) => Promise<Tokens>;
    // account_balance_dfx: (arg_0: AccountBalanceArgs) => Promise<Tokens>;
    // archives: () => Promise<Archives>;
    // decimals: () => Promise<Decimals>;
    // icrc1_balance_of: (arg_0: Account) => Promise<bigint>;
    // icrc1_decimals: () => Promise<number>;
    // icrc1_fee: () => Promise<bigint>;
    // icrc1_metadata: () => Promise<Array<[string, Value]>>;
    // icrc1_minting_account: () => Promise<[] | [Account]>;
    // icrc1_name: () => Promise<string>;
    // icrc1_supported_standards: () => Promise<Array<StandardRecord>>;
    // icrc1_symbol: () => Promise<string>;
    // icrc1_total_supply: () => Promise<bigint>;
    // icrc1_transfer: (arg_0: TransferArg) => Promise<Result>;
    // name: () => Promise<Name>;
    // query_blocks: (arg_0: GetBlocksArgs) => Promise<QueryBlocksResponse>;
    // send_dfx: (arg_0: SendArgs) => Promise<bigint>;
    // symbol: () => Promise<Symbol>;
    transfer: (arg_0: TransferArgs) => Promise<Result_1>;
    // transfer_fee: (arg_0: {}) => Promise<TransferFee>;
    // transfer_standard_stdldg: (arg_0: TransferStandardArgs) => Promise<Result_1>;
}